+++ lib/home_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fruit_panda/fruit_profiles_page.dart';
-import 'package:fruit_panda/fruit_statistics_page.dart';
+import 'package:fruit_panda/fruit_histories_page.dart';
 import 'package:fruit_panda/panda_page.dart';
 
 class HomePage extends StatelessWidget {
@@ -19,7 +19,7 @@ class HomePage extends StatelessWidget {
           children: [
             const PandaPage(),
             const FruitProfilesPage(),
-            const FruitStatisticsPage(),
+            const FruitHistoriesPage(),
             Container(color: Colors.red),
           ],
         ),
@@ -27,7 +27,7 @@ class HomePage extends StatelessWidget {
           color: Theme.of(context).colorScheme.primary,
           child: const TabBar(
             tabs: [
-              Tab(icon: Icon(Icons.looks_one), child: Text('FruitPanda')),
+              Tab(icon: Icon(Icons.looks_one), child: Text('Fruit Panda')),
               Tab(icon: Icon(Icons.looks_two), child: Text('Fruit Profiles')),
               Tab(icon: Icon(Icons.looks_3), child: Text('Fruit Statistics')),
               Tab(icon: Icon(Icons.looks_4), child: Text('Achievements')),
